+++ sdk/digitaltwins/azure-digitaltwins-core/src/samples/java/com/azure/digitaltwins/core/SyncSample.java
@@ -4,7 +4,7 @@
 package com.azure.digitaltwins.core;
 
 import com.azure.core.credential.TokenCredential;
-import com.azure.digitaltwins.core.models.DigitalTwinsGetByIdResponse;
+import com.azure.core.http.rest.PagedIterable;
 import com.azure.identity.ClientSecretCredentialBuilder;
 
 public class SyncSample
@@ -15,7 +15,12 @@ public class SyncSample
         String clientId = System.getenv("CLIENT_ID");
         String clientSecret = System.getenv("CLIENT_SECRET");
         String endpoint = System.getenv("DIGITAL_TWINS_ENDPOINT");
-        String digitalTwinId = System.getenv("DIGITAL_TWIN_ID");
+        String sourceDigitalTwinId = System.getenv("SOURCE_DIGITAL_TWIN_ID");
+        String sourceDigitalTwin = System.getenv("SOURCE_DIGITAL_TWIN");
+        String targetDigitalTwinId = System.getenv("TARGET_DIGITAL_TWIN_ID");
+        String targetDigitalTwin = System.getenv("TARGET_DIGITAL_TWIN");
+        String relationshipId = System.getenv("RELATIONSHIP_ID");
+        String relationship = System.getenv("RELATIONSHIP");
 
         TokenCredential tokenCredential = new ClientSecretCredentialBuilder()
             .tenantId(tenantId)
@@ -28,8 +33,29 @@ public class SyncSample
             .endpoint(endpoint)
             .buildClient();
 
-        DigitalTwinsGetByIdResponse syncResponse = client.getDigitalTwin(digitalTwinId);
-        Object digitalTwin = syncResponse.getValue();
-        System.out.println(digitalTwin);
+        // Create relationship on a digital twin
+        String createdRelationship = client.createRelationship(sourceDigitalTwinId, relationshipId, relationship);
+        System.out.println("Created relationship: " + createdRelationship);
+
+        // List all relationships on a digital twin
+        PagedIterable<Object> relationships = client.listRelationships(sourceDigitalTwinId, relationshipId);
+
+        // Process using the Stream interface by iterating over each page
+        relationships
+            // You can also subscribe to pages by specifying the preferred page size or the associated continuation token to start the processing from.
+            .streamByPage()
+            .forEach(page -> {
+                System.out.println("Response headers status code is " + page.getStatusCode());
+                page.getValue().forEach(item -> System.out.println("Relationship retrieved: " + item));
+            });
+
+        // Process using the Iterable interface by iterating over each page
+        relationships
+            // You can also subscribe to pages by specifying the preferred page size or the associated continuation token to start the processing from.
+            .iterableByPage()
+            .forEach(page -> {
+                System.out.println("Response headers status code is " + page.getStatusCode());
+                page.getValue().forEach(item -> System.out.println("Relationship retrieved: " + item));
+            });
     }
 }