+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/models/BlobProperties.java
@@ -57,6 +57,7 @@ public final class BlobProperties {
     private final String objectReplicationDestinationPolicyId;
     private final RehydratePriority rehydratePriority;
     private final Boolean isSealed;
+    private final OffsetDateTime expiresOn;
 
     /**
      * Constructs a {@link BlobProperties}.
@@ -173,6 +174,74 @@ public final class BlobProperties {
         final Integer committedBlockCount, final String versionId, final Boolean isCurrentVersion,
         final Long tagCount, Map<String, String> objectReplicationStatus, final String rehydratePriority,
         final Boolean isSealed) {
+        this(creationTime, lastModified, eTag, blobSize, contentType, contentMd5, contentEncoding,
+            contentDisposition, contentLanguage, cacheControl, blobSequenceNumber, blobType, leaseStatus,
+            leaseState, leaseDuration, copyId, copyStatus, copySource, copyProgress, copyCompletionTime,
+            copyStatusDescription, isServerEncrypted, isIncrementalCopy, copyDestinationSnapshot, accessTier,
+            isAccessTierInferred, archiveStatus, encryptionKeySha256, encryptionScope, accessTierChangeTime,
+            metadata, committedBlockCount, versionId, isCurrentVersion, tagCount, objectReplicationStatus,
+            rehydratePriority, isSealed, null);
+    }
+
+    /**
+     * Constructs a {@link BlobProperties}.
+     *
+     * @param creationTime Creation time of the blob.
+     * @param lastModified Datetime when the blob was last modified.
+     * @param eTag ETag of the blob.
+     * @param blobSize Size of the blob.
+     * @param contentType Content type specified for the blob.
+     * @param contentMd5 Content MD5 specified for the blob.
+     * @param contentEncoding Content encoding specified for the blob.
+     * @param contentDisposition Content disposition specified for the blob.
+     * @param contentLanguage Content language specified for the blob.
+     * @param cacheControl Cache control specified for the blob.
+     * @param blobSequenceNumber The current sequence number for a page blob, if the blob is an append or block blob
+     * pass {@code null}.
+     * @param blobType Type of the blob.
+     * @param leaseStatus Status of the lease on the blob.
+     * @param leaseState State of the lease on the blob.
+     * @param leaseDuration Type of lease on the blob.
+     * @param copyId Identifier of the last copy operation performed on the blob.
+     * @param copyStatus Status of the last copy operation performed on the blob.
+     * @param copySource Source of the last copy operation performed on the blob.
+     * @param copyProgress Progress of the last copy operation performed on the blob.
+     * @param copyCompletionTime Datetime when the last copy operation on the blob completed.
+     * @param copyStatusDescription Description of the last copy operation on the blob.
+     * @param isServerEncrypted Flag indicating if the blob's content is encrypted on the server.
+     * @param isIncrementalCopy Flag indicating if the blob was incrementally copied.
+     * @param copyDestinationSnapshot Snapshot identifier of the last incremental copy snapshot for the blob.
+     * @param accessTier Access tier of the blob.
+     * @param isAccessTierInferred Flag indicating if the access tier of the blob was inferred from properties of the
+     * blob.
+     * @param archiveStatus Archive status of the blob.
+     * @param encryptionKeySha256 SHA256 of the customer provided encryption key used to encrypt the blob on the server.
+     * @param encryptionScope The name of the encryption scope under which the blob is encrypted.
+     * @param accessTierChangeTime Datetime when the access tier of the blob last changed.
+     * @param metadata Metadata associated with the blob.
+     * @param committedBlockCount Number of blocks committed to an append blob, if the blob is a block or page blob
+     * pass {@code null}.
+     * @param versionId The version identifier of the blob.
+     * @param isCurrentVersion Flag indicating if version identifier points to current version of the blob.
+     * @param tagCount Number of tags associated with the blob.
+     * @param objectReplicationStatus The object replication status map to parse.
+     * @param rehydratePriority The rehydrate priority
+     * @param isSealed Whether or not the blob is sealed.
+     * @param expiresOn The time when the blob is going to expire.
+     */
+    public BlobProperties(final OffsetDateTime creationTime, final OffsetDateTime lastModified, final String eTag,
+        final long blobSize, final String contentType, final byte[] contentMd5, final String contentEncoding,
+        final String contentDisposition, final String contentLanguage, final String cacheControl,
+        final Long blobSequenceNumber, final BlobType blobType, final LeaseStatusType leaseStatus,
+        final LeaseStateType leaseState, final LeaseDurationType leaseDuration, final String copyId,
+        final CopyStatusType copyStatus, final String copySource, final String copyProgress,
+        final OffsetDateTime copyCompletionTime, final String copyStatusDescription, final Boolean isServerEncrypted,
+        final Boolean isIncrementalCopy, final String copyDestinationSnapshot, final AccessTier accessTier,
+        final Boolean isAccessTierInferred, final ArchiveStatus archiveStatus, final String encryptionKeySha256,
+        final String encryptionScope, final OffsetDateTime accessTierChangeTime, final Map<String, String> metadata,
+        final Integer committedBlockCount, final String versionId, final Boolean isCurrentVersion,
+        final Long tagCount, Map<String, String> objectReplicationStatus, final String rehydratePriority,
+        final Boolean isSealed, final OffsetDateTime expiresOn) {
         this.creationTime = creationTime;
         this.lastModified = lastModified;
         this.eTag = eTag;
@@ -231,6 +300,7 @@ public final class BlobProperties {
         }
         this.rehydratePriority = RehydratePriority.fromString(rehydratePriority);
         this.isSealed = isSealed;
+        this.expiresOn = expiresOn;
     }
 
 
@@ -331,6 +401,7 @@ public final class BlobProperties {
         this.objectReplicationDestinationPolicyId = objectReplicationDestinationPolicyId;
         this.rehydratePriority = null;
         this.isSealed = null;
+        this.expiresOn = null;
     }
 
     /**
@@ -618,5 +689,12 @@ public final class BlobProperties {
      */
     public Boolean isSealed() {
         return isSealed;
+    }
+
+    /**
+     * @return the time when the blob is going to expire.
+     */
+    public OffsetDateTime getExpiresOn() {
+        return expiresOn;
     }
 }